requirements.txt
PyQt5>=5.15.0
PyQt-Fluent-Widgets>=1.5.0
tushare>=1.2.0
akshare>=1.16.0
baostock>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
backtrader>=1.9.0
PyYAML>=6.0
openpyxl>=3.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyqtgraph>=0.13.0
colorlog>=6.7.0
requests>=2.31.0
cryptography>=41.0.0
pyinstaller>=6.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
easytrader>=0.20.0
